move node specific libs to separate folder
