+++ build.js
@@ -48,6 +48,9 @@ const files = {
   "dist/jacdac.cjs": "src/jacdac.ts",
   "dist/jacdac.js": "src/jacdac.ts",
   "dist/jacdac.mjs": "src/jacdac.ts",
+  "dist/jacdac-node.cjs": "src/node/jacdac-node.ts",
+  "dist/jacdac-node.js": "src/node/jacdac-node.ts",
+  "dist/jacdac-node.mjs": "src/node/jacdac-node.ts",
   "dist/p5.jacdac.js": "src/p5/p5.jacdac.ts",
   "dist/jacdac-azure-iot.mjs": "src/azure-iot/jacdac-azure-iot.ts",
   "dist/jacdac-azure-iot.cjs": "src/azure-iot/jacdac-azure-iot.ts",
