+++ src/jdom/jacdac-jdom.ts
@@ -55,9 +55,6 @@ export * from "./transport/microbit"
 export * from "./transport/workertransport"
 export * from "./transport/usb"
 export * from "./transport/webserial"
-export * from "./transport/nodewebusb"
-export * from "./transport/nodewebserialio"
-export * from "./transport/nodesocket"
 export * from "./transport/createbus"
 export * from "./error"
 export * from "./rolemanager"
